pyproject.toml
[project]
name = "imageclassifieranno"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "json5>=0.12.0",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "pyqt6>=6.9.1",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
